works with melodic & gazebo 7 (probably)

+++ include/gazebo_ros_dc_motor/gazebo_ros_dc_motor.h
@@ -104,7 +104,7 @@ namespace gazebo {
 
       void publishRotorVelocity(double m_vel);
       void publishEncoderCount(long ctr);
-      void cmdVelCallback(const std_msgs::Float32::ConstPtr& cmd_msg);
+      void cmdVelCallback(const std_msgs::Float64::ConstPtr& cmd_msg);
       void publishWheelJointState();
       void motorModelUpdate(double dt, double actual_omega, double current_torque);
       void publishEncoderCount(double m_vel, double dT);
